banking vdev Banking.APIWeb.BackofficeController

The controller containing the Backoffice operations

The operation's are executed by a logged user/pass using basic_auth.

There are four types of reports

Daily Report (see daily_report/2)

Monthly Report (see monthly_report/2)

Yearly Report (see yearly_report/2)

Total Report (see total_report/2)

Functions

daily_report(conn, _)

daily_report(Conn.t(), Map.t()) :: Plug.Conn.t()

Get Daily reports

The report will be generated as json with a list of withdrawal's, deposit's and transference's issued and received, grouped by Day, and the date representing the day.

see Banking.Backoffice.Reports.daily/0

monthly_report(conn, _)

monthly_report(Conn.t(), Map.t()) :: Plug.Conn.t()

Get Monthly reports

The report will be generated as json with a list of withdrawal's, deposit's and transference's issued and received, grouped by Month, and the date representing the Month.

see Banking.Backoffice.Reports.monthly/0

total_report(conn, _)

total_report(Conn.t(), Map.t()) :: Plug.Conn.t()

Get Total Report

The report will be generated as json with withdrawal's, deposit's and transference's issued and received, the total sum of these operations.

see Banking.Backoffice.Reports.total/0

yearly_report(conn, _)

yearly_report(Conn.t(), Map.t()) :: Plug.Conn.t()

Get Yearly reports

The report will be generated as json with a list of withdrawal's, deposit's and transference's issued and received, grouped by Year, and the date representing the year.

see Banking.Backoffice.Reports.yearly/0