banking vdev Banking.APIWeb.TransactionController

Routes implementation for Bank Account Transaction's

All Operations are executed on a private/logged API, so the Account id are in the connection. Contains five operations, and two main groups of operation:

Money input/output:

Deposit (see deposit/2)

Withdrawal (see withdrawal/2)

Transfer (see transfer/2)

Bank Account statement's:

Bank Account Statement (see statement/2)

Bank Account Balance (see balance/2)

Money format accepted in BRL format, examples: 1.000,00 or 1000 or 1000,00

Wrong way: Using dot as pennies 1000.00 Will be parsed as: 100.000,00(Representing BRL format)

Functions

balance(conn, _)

balance(Plug.Conn.t(), Map.t()) :: Plug.Conn.t()

Get Bank Account Current Balance

For more information see: Banking.Accounts.Transactions.Statement.get_balance/1

Parameters

- conn: The connection
- params: empty

deposit(conn, arg2)

deposit(Plug.Conn.t(), Map.t()) :: Plug.Conn.t()

Bank Account Deposit

Parse the amount and create a Deposit.

For more information see: Banking.Accounts.Transactions.Deposit Banking.Accounts.Utils.parse_to_decimal/1

Parameters

- conn: The connection
- params: The params provided by user containing the `amount` to be deposited

statement(conn, _)

statement(Plug.Conn.t(), Map.t()) :: Plug.Conn.t()

Get Bank Account Statement

For more information see: Banking.Accounts.Transactions.Statement.get/1

The return are parsed on Banking.APIWeb.TransactionView

Parameters

- conn: The connection
- params: empty

transfer(conn, arg2)

transfer(Plug.Conn.t(), Map.t()) :: Plug.Conn.t()

Bank Account Transference

Parse the amount and create a Transfer using the account provided as ´username´.

For more information see: Banking.Accounts.Transactions.Transfer Banking.Accounts.Utils.parse_to_decimal/1

Parameters

- conn: The connection
- params: The params provided by user containing the `amount` to be withdrawal

withdrawal(conn, arg2)

withdrawal(Plug.Conn.t(), Map.t()) :: Plug.Conn.t()

Bank Account Withdrawal

Parse the amount and create a Withdrawal.

For more information see: Banking.Accounts.Transactions.Withdrawal Banking.Accounts.Utils.parse_to_decimal/1

Parameters

- conn: The connection
- params: The params provided by user containing the `amount` to be withdrawal