banking v0.1.0 Banking.AccountManagement

Provide functions to manage accounts and users

Functions

account_from_user(user)

Get the account from a given user

Example:

iex> Banking.AccountManagement.account_from_user(user)
%Banking.AccountManagement.Account{...}

create(registration)

create(Banking.AccountManagement.Registration.t()) ::
  {:ok,
   %{
     account: Banking.AccountManagement.Account.t(),
     user: Banking.AccountManagement.User.t()
   }}
  | {:error, Ecto.Changeset.t(Banking.AccountManagement.Registration.t())}

Register a user with the following fields:

name - The FULL NAME of the user.

email

password - Must have length between 8 and 128 characters, contains a number, an upper-case, a lower-case and a symbol

birthday

document_id with document_type - The document if for the given type that could be of the following: ["cpf", "rg", "passport" or "cnpj"].

Example:

iex> attrs = %Registration{
...>     name: "User Test",
...>     email: "user@test.com",
...>     password: "veRylonGAnd$Tr0ngPasSwrd",
...>     birthdate: ~D[2000-01-01],
...>     document_id: "000.000.000-00",
...>     document_type: "cpf"
...>   }
iex> AccountManagement.create(attrs)
{:ok, %{account: %Banking.AccountManagement.Account{...}, user: %Banking.AccountManagement.User{...}}}

disable(account)

disable(Banking.AccountManagement.Account.t()) ::
  Banking.AccountManagement.Account.t()

Block a given account

Example:

iex> Banking.AccountManagement.disable(account)
%Banking.AccountManagement.Account{status: "inactive"}

enable(account)

enable(Banking.AccountManagement.Account.t()) ::
  Banking.AccountManagement.Account.t()

Enable a given account

Example:

iex> Banking.AccountManagement.enable(account)
%Banking.AccountManagement.Account{status: "active"}

get(account_id)

get(integer()) :: Banking.AccountManagement.Account.t() | nil

Get an account by his id.

Example:

iex> AccountManagement.get(1)
%Banking.AccountManagement.Account{...}

get_by_public_id(uuid)

get_by_public_id(Ecto.UUID.t()) :: Banking.AccountManagement.Account.t() | nil

Get account by the public available UUID

Example:

iex> Banking.AccountManagement.get_user_by_email("6e91802b-0f5c-43ac-bbb8-adae12ec6b62")
%Banking.AccountManagement.Account{...}

get_user_by_email(email)

get_user_by_email(String.t()) :: Banking.AccountManagement.User.t() | nil

Get user by the public available UUID

Example:

iex> Banking.AccountManagement.get_user_by_email("user@test.com")
%Banking.AccountManagement.User{...}

get_user_by_pending_email_code(email)

get_user_by_pending_email_code(String.t()) ::
  Banking.AccountManagement.User.t() | nil

Get the account by the pending email verification code

Example:

iex> code = "3Ujd31gpI6DuB8IVB2k7ahJJFKqfl0ybSa5khMQ0cVKtQFsNRWSHKEzwftpsV9T4"
iex> Banking.AccountManagement.get_user_by_pending_email_code(code)
%Banking.AccountManagement.User{...}

users_with_pending_email?(email)

users_with_pending_email?(String.t()) :: boolean()

Verifies if there is users with the given email as pending validation

Example:

iex> Banking.AccountManagement.users_with_pending_email("user@test.com")
true

validate_email(user)

validate_email(Banking.AccountManagement.User.t()) ::
  {:ok,
   %{
     account: Banking.AccountManagement.Account.t(),
     user: Banking.AccountManagement.User.t()
   }}
  | {:error, Ecto.Changeset.t()}

Validate the email by moving pending_email to the main email field.

If the user has a pending confirmation account associated then this account will be activated.

Example:

iex> Banking.AccountManagement.validate_email(user)
{:ok, %{account: Banking.AccountManagement.Account{...}, user: Banking.AccountManagement.User{...}}}

verify_password(user, password)

verify_password(Banking.AccountManagement.User.t(), String.t()) :: true | false

Compares the raw password against the hashed one presented in user

Examples:

iex> Banking.AccountManagement.verify_password(user, "correct password")
true

iex> Banking.AccountManagement.verify_password(user, "bad password")
false