banking vdev Banking.Accounts.Transactions.Deposit

The way to create a Bank Operation Deposit

Each Deposit create an Banking.Model.Balances.EventSchema.t/0 of Type.Deposit and Update the balance according with the Last Balance Event

see create/2

Functions

create(account_id, amount)

create(String.t(), Decimal.t()) ::
  {:ok, Banking.Model.Trades.TradeSchema.t()} | {:error, Ecto.Changeset.t()}

Build and write a Deposit operation on Database

This is the unique way to create a Deposit, this function creates a Banking.Model.Balances.EventSchema.t/0 and t:Banking.Model.Trades.TradesSchema.t/0 representing Trade on Account.

The account_id is provided by the Auth System and shouldn't have be provided in other way.

Parameters

- account_id: Value representing the account ID.
- amount: Decimal amount of deposit.