banking vdev Banking.Backoffice.Reports

The 'Business Logic' for reports on backoffice

Aggregates all Reports operations calling the respective function on Banking.Model.Trades.TradeLoader

Functions

daily()

daily() :: [Banking.Model.Trades.TradeLoader.t_day_report()] | []

Daily report containing a list with maps like the example above:

[%{
  day_date: ~N[2019-04-06 00:00:00.000000],
  deposit_amount: #Decimal<240000.0000>,
  transfer_issued_amount: #Decimal<-21000.0000>,
  transfer_received_amount: #Decimal<21000.0000>,
  withdrawal_amount: #Decimal<-151000.0000>
}]

see Banking.Model.Trades.TradeLoader.get_report_by_day/0 see Banking.APIWeb.BackofficeController for more information

monthly()

monthly() :: [Banking.Model.Trades.TradeLoader.t_month_report()] | []

Monthly report containing a list with maps like the example above:

[%{
  month_date: ~N[2019-03-01 00:00:00.000000],
  deposit_amount: #Decimal<240000.0000>,
  transfer_issued_amount: #Decimal<-21000.0000>,
  transfer_received_amount: #Decimal<21000.0000>,
  withdrawal_amount: #Decimal<-151000.0000>
}]

see Banking.Model.Trades.TradeLoader.get_report_by_month/0 see Banking.APIWeb.BackofficeController for more information

total()

total() :: Banking.Model.Trades.TradeLoader.t_total()

Yearly report containing a Map like the example above:

%{
deposit_amount: #Decimal<240000.0000>,
transfer_issued_amount: #Decimal<-21000.0000>,
transfer_received_amount: #Decimal<21000.0000>,
withdrawal_amount: #Decimal<-151000.0000>
}

see Banking.Model.Trades.TradeLoader.get_report_total/0 see Banking.APIWeb.BackofficeController for more information

yearly()

yearly() :: [Banking.Model.Trades.TradeLoader.t_year_report()] | []

Yearly report containing a list with maps like the example above:

[%{
  year_date: ~N[2018-01-01 00:00:00.000000],
  deposit_amount: #Decimal<240000.0000>,
  transfer_issued_amount: #Decimal<-21000.0000>,
  transfer_received_amount: #Decimal<21000.0000>,
  withdrawal_amount: #Decimal<-151000.0000>
}]

see Banking.Model.Trades.TradeLoader.get_report_by_month/0 see Banking.APIWeb.BackofficeController for more information