banking v0.1.0 Banking.Bank

Processes all the financial operations and make an immutable register them as transactions.

Functions

add_bonus(target, amount)

Add funds to a given account

Example:

iex> Bank.withdrawal(account, 10) {:ok, %{

bank_transaction: %Banking.Bank.Transaction{
__meta__: #Ecto.Schema.Metadata<:loaded, "bank_transactions">,
amount: #Decimal<1000>,
id: 3,
inserted_at: ~N[2019-11-28 04:40:42],
public_id: "9086a2b5-7f21-494b-9d83-28c2a10d0432",
source_id: 1,
target_id: 4,
updated_at: ~N[2019-11-28 04:40:42]

}}

transfer(source, target, amount)

Transfer money from the source account to the target.

Example:

iex> Bank.transfer(source_account, target_account, 100) {:ok, %{

bank_transaction: %Banking.Bank.Transaction{
  __meta__: #Ecto.Schema.Metadata<:loaded, "bank_transactions">,
  amount: #Decimal<100>,
  id: 1,
  inserted_at: ~N[2019-11-28 04:36:24],
  public_id: "54469e74-79da-468c-aafc-6a0bbfb4af8e",
  source_id: 4,
  target_id: 5,
  updated_at: ~N[2019-11-28 04:36:24]
}}

withdrawal(source, amount)

Withdrawal money from a given account

Example:

iex> Bank.withdrawal(account, 10) {:ok, %{

bank_transaction: %Banking.Bank.Transaction{
__meta__: #Ecto.Schema.Metadata<:loaded, "bank_transactions">,
amount: #Decimal<10>,
id: 2,
inserted_at: ~N[2019-11-28 04:37:54],
public_id: "c5f3e493-11c4-4703-82c7-c001f7294619",
source_id: 4,
target_id: 2,
updated_at: ~N[2019-11-28 04:37:54]

}}