banking v0.1.0 Banking.Bank.KPI

Provides many different indicators about the bank operations

Functions

transactions_sum()

Calculates the sum of amount from all transactions in history

Example:

iex> Banking.Bank.KPI.transactions_sum()
#Decimal<1000.0000>

transactions_sum_current_month()

Calculates the sum of amount from all transactions created this month

Example:

iex> Banking.Bank.KPI.transactions_sum_current_month()
#Decimal<1000.0000>

transactions_sum_current_year()

Calculates the sum of amount from all transactions created this year

Example:

iex> Banking.Bank.KPI.transactions_sum_current_year()
#Decimal<1000.0000>

transactions_sum_past_days(days)

Calculates the sum of amount from all transactions created in the past n days

Example:

iex> Banking.Bank.KPI.transactions_sum_past_days()
#Decimal<1000.0000>

transactions_sum_previous_month()

Calculates the sum of amount from all transactions created past month

Example:

iex> Banking.Bank.KPI.transactions_sum_previous_month()
#Decimal<1000.0000>

transactions_sum_previous_year()

Calculates the sum of amount from all transactions created past year

Example:

iex> Banking.Bank.KPI.transactions_sum_previous_year()
#Decimal<1000.0000>