banking v0.1.0 Banking.Bank.KPI.AmountVolume

Reports the total amount volume that was transferred or withdraw

Functions

current_month()

The sum of amount from all transactions created this month

current_year()

The sum of amount from all transactions created this year

past_days(days)

The sum of amount from all transactions created in the past n days

previous_month()

The sum of amount from all transactions created past month

previous_year()

The sum of amount from all transactions created past year

total()

The sum of amount from all transactions in history