banking vdev Banking.Model.Balances.EventSchema

The Balances of an bank account

This schema control every fund added or removed of an account, the final balance is represented by balance field

Each event have the previous event t/0 as a parent represented as parent

The type is an Enum Banking.Model.Enums.BankOperationEnum.t/0 corresponding about Bank Account operation, the values accepted are "DEPOSIT", "TRANSFER_ISSUED", "TRANSFER_RECEIVED" and "WITHDRAWAL"

Types

t()

t() :: %Banking.Model.Balances.EventSchema{
  __meta__: term(),
  account: term(),
  account_id: term(),
  balance: term(),
  id: term(),
  inserted_at: term(),
  parent: term(),
  parent_id: term(),
  quantity_moved: term(),
  type: term(),
  updated_at: term()
}

The Banking.Model.Balances.EventSchema struct.

Functions

changeset(event \\ __MODULE__, params \\ %{})

changeset(t(), Map.t()) :: Changeset.t()

The changeset for t/0

Responsible also to set the sign of quantity_moved by Banking.Model.Enums.BankOperationEnum.t/0

Do not accept the balance as params the balance is calculate by the changeset/2 using the previous t/0(:parent) using the balance of parent and the current quantity_moved

Examples

iex> EventSchema.changeset %EventSchema{}, %{type: Type.Deposit, quantity_moved: 10, account_id: account.id}
#Ecto.Changeset<
action: nil,
changes: %{
  account_id: 2,
  balance: #Decimal<10>,
  quantity_moved: #Decimal<10>,
  type: Banking.Model.Enums.BankOperationEnum.Type.Deposit
},
errors: [],
data: #Banking.Model.Balances.EventSchema<>,
valid?: true
>