banking vdev Banking.Model.Trades.TradeSchema

Represent a Trade

The trade is each Trade event, like Deposit, Withdrawal, Transfer received and issued and is backed on Banking.Model.Balances.EventSchema.t/0.

The fields type, balance, quantity are virtual's field, necessary to build the reports, loaded from balance_event

Types

t()

t() :: %Banking.Model.Trades.TradeSchema{
  __meta__: term(),
  account: term(),
  account_id: term(),
  balance: term(),
  balance_event: term(),
  balance_event_id: term(),
  id: term(),
  inserted_at: term(),
  quantity: term(),
  transfer_account: term(),
  transfer_account_id: term(),
  type: term(),
  updated_at: term()
}

The Banking.Model.Trades.TradeSchema struct.

Functions

changeset(event \\ __MODULE__, params \\ %{}, event_type)

changeset(t(), Map.t(), String.t()) :: Changeset.t()

The changeset for t/0

The values are created based on Balance Event, the event_type param is the :type on Banking.Model.Balances.EventSchema.t/0 and is used to support the transfer_account_id validation (that may be required or not, depending of the type of transaction)

Examples

iex> TradeSchema.changeset(%TradeSchema{}, %{}, type: "DEPOSIT")
#Ecto.Changeset<
action: nil,
changes: %{type: Banking.Model.Enums.BankOperationEnum.Type.Deposit},
errors: [
  account_id: {"can't be blank", [validation: :required]},
  balance_event_id: {"can't be blank", [validation: :required]}
],
data: #Banking.Model.Trades.TradeSchema<>,
valid?: false
>