banking v0.1.0 Banking.Session

Handles the user session by tracking the IP and device used.

Here you can put extra behaviors like controlling simultaneous sessions.

Types

jwt()

jwt() :: String.t()

session_opts()

session_opts() :: %{device_type: String.t(), device_id: String.t()} | %{}

Functions

create(user, ip_address, opts \\ %{})

create(
  Banking.AccountManagement.User.t(),
  Banking.Session.SessionSchema.ip(),
  session_opts()
) :: {:ok, Banking.Session.SessionSchema.t()} | {:error, Ecto.Changeset.t()}

Creates a session for a given user

Example:

iex> Banking.Session.create(user, "127.0.0.1", device_type: "android", device_id: "1231...")
{:ok, %Banking.Session.SessionSchema{...}}

create_jwt(user, ip_address, opts \\ %{})

create_jwt(
  Banking.AccountManagement.User.t(),
  Banking.Session.SessionSchema.ip(),
  session_opts()
) :: {:ok, jwt(), map()}

Creates a session for a given user and returns token and claims

Example:

iex> Banking.Session.create_jwt(user, "127.0.0.1", device_type: "android", device_id: "1231...")
{:ok, "jwt...",  %{
  "aud" => "banking",
  "exp" => 1576646266,
  "iat" => 1576473466,
  "iss" => "banking",
  "jti" => "40bf6667-dca9-42b2-92ab-461cfd9c7a67",
  "nbf" => 1576473465,
  "sub" => "3942",
  "typ" => "access"
}}

get(session_id)

get(integer()) :: Banking.Session.SessionSchema.t() | nil

Get session by id

Example:

iex> Banking.Session.get(1)
%Banking.Session.SessionSchema{...}

iex> Banking.Session.get(1000)
nil

to_jwt(session)

to_jwt(Banking.Session.SessionSchema.t()) :: {:ok, jwt(), map()}

Generates Jason Web Token and claims from an existing session

Example:

iex> Banking.Session.to_jwt(session)
{:ok, "jwt...",  %{
  "aud" => "banking",
  "exp" => 1576646266,
  "iat" => 1576473466,
  "iss" => "banking",
  "jti" => "40bf6667-dca9-42b2-92ab-461cfd9c7a67",
  "nbf" => 1576473465,
  "sub" => "3942",
  "typ" => "access"
}}