banking v0.1.0 Banking.AccountManagement

Works as a public API to handle all the operations regarding user registration and account administration.

Functions

create(registration)

create(Banking.AccountManagement.Registration.t()) ::
  {:ok,
   %{
     account: Banking.AccountManagement.Account.t(),
     user: Banking.AccountManagement.User.t()
   }}
  | {:error, Ecto.Changeset.t(Banking.AccountManagement.Registration.t())}

Register a user with the following fields:

name - The FULL NAME of the user.

email

password - Must have length between 8 and 128 characters, contains a number, an upper-case, a lower-case and a symbol

birthday

document_id with document_type - The document if for the given type that could be of the following: cpf, rg, passport or cnpj.

disable(account)

disable(Banking.AccountManagement.Account.t()) ::
  Banking.AccountManagement.Account.t()

Block a given account

enable(account)

enable(Banking.AccountManagement.Account.t()) ::
  Banking.AccountManagement.Account.t()

Enable a given account

get(account_id)

get(integer()) :: Banking.AccountManagement.Account.t() | nil

Get an account by his id.

get_by_public_id(uuid)

get_by_public_id(Ecto.UUID.t()) :: Banking.AccountManagement.Account.t() | nil

Get the account by the public available UUID

validate_email(user)

validate_email(Banking.AccountManagement.User.t()) ::
  {:ok,
   %{
     account: Banking.AccountManagement.Account.t(),
     user: Banking.AccountManagement.User.t()
   }}
  | {:error, <<_::152>>}

Validate the email by moving pending_email to the main email field.

If the user has a pending confirmation account associated then this account will be activated.

verify_password(user, password)

verify_password(Banking.AccountManagement.User.t(), String.t()) :: true | false

Verifies the hashed password from the user against the raw password